+++ src/builtins/pwd.c
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 16:54:05 by emadriga          #+#    #+#             */
-/*   Updated: 2021/11/28 11:40:09 by emadriga         ###   ########.fr       */
+/*   Updated: 2021/12/11 10:38:46 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 */
 void	ft_pwd(t_str **env_list, char **argv)
 {
-	char	cwd[PATH_MAX - 1];
+	char	cwd[PATH_MAX];
 
 	(void)env_list;
 	if (argv[1] != NULL && argv[1][0] != '|' && argv[1][0] != '<' \
@@ -29,9 +29,7 @@ void	ft_pwd(t_str **env_list, char **argv)
 		log_error(ENV_NOT_OPTIONS_ALLOWED, 1);
 	else
 	{	
-		if (getcwd(cwd, sizeof(cwd)) == NULL)
-			perror("getcwd() error");
-		else
-			printf("%s\n", cwd);
+		getcwd(cwd, PATH_MAX);
+		printf("%s\n", cwd);
 	}
 }