+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2021-12-11 - Edu
+
+### Fixed
+- **builtins/cd.c** fix log error 'No such file or directory'
+- **builtins/cd.c** & **builtins/pwd.c** getcwd doesn't check if current working directory exist anymore to emulate bash work
+
 ## 2021-11-30 - Jorge
 
 ### Changed